pyproject.toml
[project]
name = "workgarden"
version = "0.1.0"
description = "CLI tool for managing git worktrees with Docker Compose and Claude Code support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18.0",
]

[project.scripts]
wg = "workgarden.cli.app:app"
workgarden = "workgarden.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/workgarden"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
